+++ flake.nix
@@ -13,7 +13,7 @@
 
         repoctx = pkgs.rustPlatform.buildRustPackage {
           pname = "repoctx";
-          version = "0.11.5";
+          version = "0.11.6";
           src = ./.;
 
           cargoLock = {